修复断点续传的bug

+++ Tiger/app/src/main/java/com/mn/tiger/download/FileSizeDownloadFileChecker.java
@@ -20,7 +20,10 @@ public class FileSizeDownloadFileChecker implements IDownloadFileChecker
 			}
 			else
 			{
-				file.delete();
+				if(!downloader.isBreakPoints())
+				{
+					file.delete();
+				}
 				return false;
 			}
 		}
